feat:17.抽奖活动流水入库

+++ imwj-dev-tech-big-market/imwj-dev-tech-big-market-domain/src/main/java/com/imwj/big/market/domain/activity/model/entity/ActivityCountEntity.java
@@ -24,17 +24,17 @@ public class ActivityCountEntity {
     /**
      * 总次数
      */
-    private Integer totalCount;
+    private Long totalCount;
 
     /**
      * 日次数
      */
-    private Integer dayCount;
+    private Long dayCount;
 
     /**
      * 月次数
      */
-    private Integer monthCount;
+    private Long monthCount;
 
 }
